chore(rollup): merge the dev/prod config
